+++ AmazonChimeSDK/AmazonChimeSDK/analytics/EventName.swift
@@ -10,8 +10,12 @@ import Foundation
 
 /// `EventName` represent some major event that could help builders to analyze the data
 @objc public enum EventName: Int, CaseIterable, CustomStringConvertible {
+    /// The microphone was selected.
+    case audioInputSelected
     /// The microphone selection or access failed
     case audioInputFailed
+    /// The camera was selected.
+    case videoInputSelected
     /// The camera selection or access failed
     case videoInputFailed
     /// The meeting will start
@@ -42,13 +46,33 @@ import Foundation
     case appStateChanged
     /// The application memory is low
     case appMemoryLow
+    /// Voice focus enabled
+    case voiceFocusEnabled
+    /// Voice focus disabled
+    case voiceFocusDisabled
+    /// Failed to enable voice focus
+    case voiceFocusEnableFailed
+    /// Failed to disable voice focus
+    case voiceFocusDisableFailed
+    /// Audio interruption began
+    case audioInterruptionBegan
+    /// Audio interruption ended
+    case audioInterruptionEnded
+    /// Video interruption began
+    case videoInterruptionBegan
+    /// Video interruption ended
+    case videoInterruptionEnded
     // unknown
     case unknown
 
     public var description: String {
         switch self {
+        case .audioInputSelected:
+            return "audioInputSelected"
         case .audioInputFailed:
             return "audioInputFailed"
+        case .videoInputSelected:
+            return "videoInputSelected"
         case .videoInputFailed:
             return "videoInputFailed"
         case .meetingStartRequested:
@@ -79,6 +103,22 @@ import Foundation
             return "appStateChanged"
         case .appMemoryLow:
             return "appMemoryLow"
+        case .voiceFocusEnabled:
+            return "voiceFocusEnabled"
+        case .voiceFocusDisabled:
+            return "voiceFocusDisabled"
+        case .voiceFocusEnableFailed:
+            return "voiceFocusEnableFailed"
+        case .voiceFocusDisableFailed:
+            return "voiceFocusDisableFailed"
+        case .audioInterruptionBegan:
+            return "audioInterruptionBegan"
+        case .audioInterruptionEnded:
+            return "audioInterruptionEnded"
+        case .videoInterruptionBegan:
+            return "videoInterruptionBegan"
+        case .videoInterruptionEnded:
+            return "videoInterruptionEnded"
         case .unknown:
             return "unknown"
         }
@@ -86,8 +126,12 @@ import Foundation
 
     static func toEventName(name: String) -> EventName {
         switch name {
+        case "audioInputSelected":
+            return .audioInputSelected
         case "audioInputFailed":
             return .audioInputFailed
+        case "videoInputSelected":
+            return .videoInputSelected
         case "videoInputFailed":
             return .videoInputFailed
         case "meetingStartRequested":
@@ -118,6 +162,22 @@ import Foundation
             return .appStateChanged
         case "appMemoryLow":
             return .appMemoryLow
+        case "voiceFocusEnabled":
+            return .voiceFocusEnabled
+        case "voiceFocusDisabled":
+            return .voiceFocusDisabled
+        case "voiceFocusEnableFailed":
+            return .voiceFocusEnableFailed
+        case "voiceFocusDisableFailed":
+            return .voiceFocusDisableFailed
+        case "audioInterruptionBegan":
+            return .audioInterruptionBegan
+        case "audioInterruptionEnded":
+            return .audioInterruptionEnded
+        case "videoInterruptionBegan":
+            return .videoInterruptionBegan
+        case "videoInterruptionEnded":
+            return .videoInterruptionEnded
         default:
             return .unknown
         }